优化id储存，优化字符串转u64

+++ 1h/mywifi.h
@@ -30,7 +30,7 @@ extern "C"
     short file_save_wifidata(char *WIFI_ssid, char *WIFI_password, const char *wifi_ssid_pw_file);
     short file_delete(const char *file_name);
     short get_wifi(char *WIFI_ssid, char *WIFI_password, const char *wifi_ssid_pw_file);
-    short tcp_server_get_wifi_data(char *WIFI_ssid, char *WIFI_password, unsigned long &UID, uint32_t CHIP_ID, const char *wifi_ssid_pw_file);
+    short tcp_server_get_wifi_data(char *WIFI_ssid, char *WIFI_password, unsigned long long &UID, uint32_t CHIP_ID, const char *wifi_ssid_pw_file);
     void clear_wifi_data(const char *file_name);
 }
 #endif
